+++ far/src/vmenu.cpp
@@ -1105,8 +1105,8 @@ int VMenu::ProcessKey(FarKey Key)
 		}
 	}
 
-	if (!(((unsigned int)Key >= KEY_MACRO_BASE && (unsigned int)Key <= KEY_MACRO_ENDBASE)
-				|| ((unsigned int)Key >= KEY_OP_BASE && (unsigned int)Key <= KEY_OP_ENDBASE))) {
+	if (!((Key >= KEY_MACRO_BASE && Key <= KEY_MACRO_ENDBASE)
+				|| (Key >= KEY_OP_BASE && Key <= KEY_OP_ENDBASE))) {
 		DWORD S = Key & (KEY_CTRL | KEY_ALT | KEY_SHIFT | KEY_RCTRL | KEY_RALT);
 		DWORD K = Key & (~(KEY_CTRL | KEY_ALT | KEY_SHIFT | KEY_RCTRL | KEY_RALT));
